[WASM] ListViewBase - refactor to use shared code
Remove references to NativePanel from shared class since it isn't applicable in WebAssembly.
