+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/ListViewBase.cs
@@ -1,4 +1,4 @@
-﻿#if !NET46 && !NETSTANDARD2_0 && !__MACOS__
+﻿#if !NET46 && !__MACOS__
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -33,7 +33,6 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class ListViewBase : Selector
 	{
-		internal NativeListViewBase NativePanel { get { return InternalItemsPanelRoot as NativeListViewBase; } }
 		internal ScrollViewer ScrollViewer { get; private set; }
 
 		/// <summary>
@@ -50,27 +49,6 @@ namespace Windows.UI.Xaml.Controls
 		/// Whether an incremental data loading request is currently under way.
 		/// </summary>
 		private bool _isIncrementalLoadingInFlight;
-		/// <summary>
-		/// The number of currently visible items, ie a 'page' from the point of view of incremental data loading.
-		/// </summary>
-		private int PageSize
-		{
-			get
-			{
-				if (NativePanel == null)
-				{
-					// Not supported
-					return 0;
-				}
-				var lastVisibleIndex = NativePanel.NativeLayout.LastVisibleIndex;
-				var firstVisibleIndex = NativePanel.NativeLayout.FirstVisibleIndex;
-				if (lastVisibleIndex == -1)
-				{
-					return 0;
-				}
-				return lastVisibleIndex + 1 - firstVisibleIndex;
-			}
-		}
 
 		protected ListViewBase()
 		{
@@ -302,33 +280,7 @@ namespace Windows.UI.Xaml.Controls
 
 			ScrollViewer = this.GetTemplateChild("ScrollViewer") as ScrollViewer;
 
-			// NativePanel may not exist if we're using a non-virtualizing ItemsPanel.
-			if (NativePanel != null)
-			{
-				NativePanel.XamlParent = this;
-				// Propagate the DataContext manually, since ItemsPanelRoot isn't really part of the visual tree
-				ItemsPanelRoot.SetValue(DataContextProperty, DataContext, DependencyPropertyValuePrecedences.Inheritance);
-				OnApplyTemplatePartial();
-
-				if (ScrollViewer?.Style?.Precedence == DependencyPropertyValuePrecedences.ImplicitStyle)
-				{
-					throw new InvalidOperationException($"Performance hit: {this} is using a ScrollViewer in its template with a default style, which would break virtualization. A Style containing {nameof(ListViewBaseScrollContentPresenter)} must be used.");
-				}
-
-				if (ScrollViewer != null)
-				{
-					NativePanel.HorizontalScrollBarVisibility = ScrollViewer.HorizontalScrollBarVisibility;
-					NativePanel.VerticalScrollBarVisibility = ScrollViewer.VerticalScrollBarVisibility;
-				}
-			}
-			else
-			{
-				if (ScrollViewer?.Style == Uno.UI.GlobalStaticResources.ListViewBaseScrollViewerStyle)
-				{
-					// We're not using NativeListViewBase so we need a 'real' ScrollViewer
-					ScrollViewer.Style = Uno.UI.GlobalStaticResources.DefaultScrollViewerStyle;
-				}
-			}
+			OnApplyTemplatePartial();
 		}
 		partial void OnApplyTemplatePartial();
 
@@ -346,31 +298,10 @@ namespace Windows.UI.Xaml.Controls
 				ApplyMultiSelectState(item);
 			}
 
-			foreach (var item in (NativePanel?.CachedItemViews).Safe())
-			{
-				ApplyMultiSelectState(item);
-			}
+			ApplyMultiSelectStateToCachedItems();
 		}
 
-		protected override _View ResolveInternalItemsPanel(_View itemsPanel)
-		{
-			// If the items panel is a virtualizing panel, we substitute it with NativeListViewBase
-			var virtualizingPanel = itemsPanel as IVirtualizingPanel;
-			if (virtualizingPanel != null)
-			{
-				var internalPanel = new NativeListViewBase();
-				var layouter = virtualizingPanel.GetLayouter();
-				PrepareNativeLayout(layouter);
-				internalPanel.NativeLayout = layouter;
-				internalPanel.BindToEquivalentProperty(virtualizingPanel, "Background");
-				return internalPanel;
-			}
-			else
-			{
-				// Otherwise act as a normal ItemsControl
-				return base.ResolveInternalItemsPanel(itemsPanel);
-			}
-		}
+		partial void ApplyMultiSelectStateToCachedItems();
 
 		partial void PrepareNativeLayout(VirtualizingPanelLayout layout);
 
@@ -784,7 +715,7 @@ namespace Windows.UI.Xaml.Controls
 			// unfilled viewport because PageSize doesn't return the 'potential' number of visible items.
 			if (result.Count > 0)
 			{
-				TryLoadMoreItems(NativePanel.NativeLayout.LastVisibleIndex);
+				TryLoadMoreItems();
 			}
 		}
 	}
