+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/ListViewBase.netstd.cs
@@ -7,69 +7,31 @@ using Windows.UI.Xaml.Data;
 
 namespace Windows.UI.Xaml.Controls
 {
-	public partial class ListViewBase : Selector
+	public partial class ListViewBase
 	{
-		// TODO: This is copied from ListViewBase.cs (shared).
 
-		public event ItemClickEventHandler ItemClick;
+		private int PageSize => throw new NotImplementedException();
 
-		public IList<object> SelectedItems { get; } = new List<object>();
-
-		internal override void OnItemClicked(int clickedIndex)
+		private void Refresh()
 		{
-			var item = ItemFromIndex(clickedIndex);
-			if (IsItemClickEnabled)
-			{
-				ItemClick?.Invoke(this, new ItemClickEventArgs { ClickedItem = item });
-			}
-
-			//Handle selection
-			switch (SelectionMode)
-			{
-				case ListViewSelectionMode.None:
-					break;
-				case ListViewSelectionMode.Single:
-					if (ItemsSource is ICollectionView collectionView)
-					{
-						//NOTE: Windows seems to call MoveCurrentTo(item); we set position instead to have expected behavior when you have duplicate items in the list.
-						collectionView.MoveCurrentToPosition(clickedIndex);
-
-						// The CollectionView may have intercepted the change
-						clickedIndex = collectionView.CurrentPosition;
-
-					}
-
-					SelectedIndex = clickedIndex;
-					SelectedItem = item;
-					break;
-				case ListViewSelectionMode.Multiple:
-				case ListViewSelectionMode.Extended:
-					HandleMultipleSelection(clickedIndex, item);
-					break;
-			}
+			//TODO
 		}
 
-		private void HandleMultipleSelection(int clickedIndex, object item)
-		{
-			if (!SelectedItems.Contains(item))
-			{
-				SelectedItems.Add(item);
-				SetSelectedState(clickedIndex, true);
-			}
-			else
-			{
-				SelectedItems.Remove(item);
-				SetSelectedState(clickedIndex, false);
-			}
-		}
+		private void AddItems(int firstItem, int count, int section) => throw new NotImplementedException();
 
-		private void SetSelectedState(int clickedIndex, bool selected)
+		private void RemoveItems(int firstItem, int count, int section) => throw new NotImplementedException();
+
+		private void AddGroup(int groupIndexInView) => throw new NotImplementedException();
+
+		private void RemoveGroup(int groupIndexInView) => throw new NotImplementedException();
+
+		private void ReplaceGroup(int groupIndexInView) => throw new NotImplementedException();
+
+		private ContentControl ContainerFromGroupIndex(int groupIndex) => throw new NotImplementedException();
+
+		private void TryLoadMoreItems()
 		{
-			var selectorItem = ContainerFromIndex(clickedIndex) as SelectorItem;
-			if (selectorItem != null)
-			{
-				selectorItem.IsSelected = selected;
-			}
+			//TODO: ISupportIncrementalLoading
 		}
 	}
 }
